requirements.txt
Flask>=3.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
itsdangerous>=2.2.0
click>=8.1.0
werkzeug>=3.0.0
pytest>=8.0.0

# LangChain + OpenAI
langchain>=0.2.0
langchain-core>=0.2.22
langchain-openai>=0.1.7
openai>=1.40.0
Flask>=3.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
itsdangerous>=2.2.0
click>=8.1.0
werkzeug>=3.0.0
pytest>=8.0.0

# LangChain + OpenAI
langchain>=0.2.0
langchain-core>=0.2.22
langchain-openai>=0.1.7
openai>=1.40.0
huggingface_hub>=0.24.0


## Jupyter Notebook
notebook==6.1.5
